requirements.txt
alembic==1.0.6
Flask==1.0.2
Flask-Cors==3.0.7
Flask-Migrate==2.3.1
Flask-SQLAlchemy==2.3.2
gunicorn==19.7.1
itsdangerous==1.1.0
Jinja2==2.10
keystoneauth1==3.11.2
keystonemiddleware==5.3.0
mysqlclient==1.4.1
PasteDeploy==2.0.1
PyMySQL==0.9.3
python-keystoneclient==3.18.0
requests==2.18.4
SQLAlchemy==1.2.16
Werkzeug==0.14.1
